Refactor Monitor Classes to use REST API
Since the REST API needs to be stable for the GUI in
the javascript, and a client facade is required for
the development of an easy-to-use CLI app. It makes
little sense to waste more time getting the SQL Sessions
to work inside the worker threads and not just have them
talk to the REST API that has all the session-management
covered anyways.